+++ CHANGELOG.md
@@ -6,12 +6,13 @@ Another round of components have been updated with improved accessibility.
 
 ### Features
 
-- `breadcrumb-item` - New `current` prop marks the current page.
+- `breadcrumb-item` - A new `current` prop marks the current page.
 - `form-actions` - A new `label` slot allows a set of actions to have a unique accessible label.
 - `form-wrapper` - A new `actions-label` slot allows a set of actions to have a unique accessible label.
 - `link-tag` - A new `external-suffix` slot allows for custom external text when a link is `external`.
 - `progress-bar` - A new `variant` prop allows using as either a progress bar or a `meter`.
 - `progress-bar` - Now allows passing `null` as the `current` value to show an indeterminate progress bar.
+- `data-table` - New `search-introduction` and `search-help` slots allow passing introduction and help content to the search field.
 
 ## 1.9.0